implement search by toys

+++ src/page/myToys/MyToys.jsx
@@ -1,6 +1,7 @@
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../context/AuthProvider";
 import MyToysTable from "./MyToysTable";
+import Swal from "sweetalert2";
 
 const MyToys = () => {
 
@@ -23,18 +24,32 @@ const MyToys = () => {
 
     // handlerDeleteToy
     const handlerDeleteToy = _id => {
-        console.log(_id);
 
-        fetch(`http://localhost:5000/my-toys${_id}`, {
-            method: "DELETE"
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
         })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data);
+            .then(result => {
+                if (result.isConfirmed) {
+                    fetch(`http://localhost:5000/my-toys${_id}`, {
+                        method: "DELETE"
+                    })
+                        .then(res => res.json())
+                        .then(data => {
+                            console.log(data);
 
-                const remaining = mineToys.filter(toy => toy._id !== _id)
-                setMineToys(remaining)
+                            const remaining = mineToys.filter(toy => toy._id !== _id)
+                            setMineToys(remaining)
+                        })
+                }
             })
+
+
     }
 
     return (
